+++ apps/designmanual-frontend/app/features/portable-text/components/ImageBlock.tsx
@@ -484,7 +484,6 @@ const GridLayout = ({
   imageCount,
   captionAndCredits,
 }: GridLayoutProps) => {
-  // eslint-disable-next-line no-nested-ternary
   const rowsMobile = imageCount <= 2 ? 2 : imageCount <= 3 ? 24 : 32;
   const rowOthers = imageCount <= 5 ? 2 : 3;
 
@@ -541,13 +540,17 @@ function calculateAspectRatio(imageCount: number, deviceType: string) {
     case "desktop": {
       switch (imageCount) {
         case 1: {
-          aspectRatio = "16/8";
+          aspectRatio = "16/9";
 
           break;
         }
         case 2:
+        case 4: {
+          aspectRatio = "1/1";
+
+          break;
+        }
         case 3:
-        case 4:
         case 5: {
           aspectRatio = "2/1";
           break;