+++ apps/designmanual-frontend/app/root/layout/Footer.tsx
@@ -20,7 +20,7 @@ export const Footer = () => {
     <Flex
       as="footer"
       justifyContent="space-between"
-      backgroundColor={"surface.secondary"}
+      backgroundColor="surface.secondary"
       className="dark"
       alignItems="center"
       gap="4"
@@ -29,10 +29,11 @@ export const Footer = () => {
       borderTop="1px solid"
       borderColor={borderColor}
       paddingX={[3, null, 6, 4, 8]}
-      width={"100%"}
+      width="100%"
       marginX="auto"
-      minHeight={"12rem"}
+      minHeight="12rem"
       flexDirection={["column", null, "row"]}
+      zIndex="banner"
     >
       <Box alignSelf="start" marginTop="4">
         <Link to="/" aria-label="Go to the front page">
@@ -46,16 +47,16 @@ export const Footer = () => {
 
       <Flex
         flexDirection="row"
-        justifyContent={"center"}
+        justifyContent="center"
         alignItems="start"
         gap={2}
         flexWrap="wrap"
         flexGrow={1}
       >
         <Flex
-          flexDirection={"column"}
+          flexDirection="column"
           gap="6"
-          justifyContent={"center"}
+          justifyContent="center"
           alignItems={["start", null, "center"]}
         >
           <Flex
@@ -68,12 +69,14 @@ export const Footer = () => {
                 flexDirection="column"
                 key={item._key}
                 textAlign={["left", null, "center"]}
+                color="text.inverted"
               >
                 <Text
                   as="h6"
                   textAlign={["left", null, "center"]}
-                  fontWeight={"bold"}
+                  fontWeight="bold"
                   variant="xs"
+                  color="text"
                 >
                   {item.title}
                 </Text>
@@ -81,9 +84,9 @@ export const Footer = () => {
                 <Box
                   textAlign={["left", null, "center"]}
                   fontSize="xs"
+                  color="text.inverted"
                   css={{
                     "& a": {
-                      color: "inherit",
                       textDecoration: "underline",
                     },
                   }}