+++ apps/designmanual-frontend/app/routes/_base/content-menu/ContentMenu.tsx
@@ -50,10 +50,10 @@ export const ContentMenu = forwardRef<
   const allSections =
     useRouteLoaderData("root")?.initialSanityData?.siteSettings?.topMenu || [];
 
-  const isProd = useRouteLoaderData("root")?.env === "prod";
+  const isProduction = useRouteLoaderData("root")?.env === "prod";
 
   const sections = allSections.filter((s: Section) => {
-    if (isProd && s.slug.current.includes("identitet")) {
+    if (isProduction && s.slug.current.includes("identitet")) {
       return false;
     }
     return true;
@@ -79,7 +79,7 @@ export const ContentMenu = forwardRef<
 
   return (
     <React.Fragment key="content-menu">
-      <Flex flexDirection={"column"} display={["flex", null, null, "none"]}>
+      <Flex flexDirection="column" display={["flex", null, null, "none"]}>
         {sections &&
           sections.length > 7 &&
           sections.map((section: Section) => (
@@ -101,7 +101,7 @@ export const ContentMenu = forwardRef<
         variant="ghost"
         collapsible
         defaultValue={expanded}
-        onValueChange={(e) => setExpanded(e.value)}
+        onValueChange={(details) => setExpanded(details.value)}
         key={refreshKey}
         display={["none", null, null, "block"]}
       >
@@ -119,11 +119,11 @@ export const ContentMenu = forwardRef<
                 url={handleExternalMenu(item.link, isPreview)}
                 isTopMenu={true}
                 ref={index === 0 ? ref : null}
-                fontWeight={"bold"}
+                fontWeight="bold"
                 fontSize={["desktop.xs", null, "desktop.sm"]}
                 paddingX="3"
                 paddingY="2"
-                borderRadius={"sm"}
+                borderRadius="sm"
                 onClick={() => {
                   setExpanded([item.link]);
                   handleRefresh();
@@ -302,7 +302,7 @@ const MobileMenu = forwardRef(
                       textAlign="left"
                       paddingLeft={6}
                       marginBottom={2}
-                      fontWeight={"bold"}
+                      fontWeight="bold"
                     >
                       <Link
                         to={handleExternalMenu(item?.link ?? "/", isPreview)}