+++ packages/spor-react/src/input/PhoneNumberInput.tsx
@@ -99,8 +99,8 @@ export const PhoneNumberInput = forwardRef<
         value={value.nationalNumber}
         invalid={invalid}
         errorText={errorText}
-        onChange={(e) => {
-          const target = e.target as HTMLInputElement;
+        onChange={(event) => {
+          const target = event.target as HTMLInputElement;
           // Removes everything but numbers, spaces and dashes
           const strippedValue = target.value.replaceAll(/[^\d\s-]/g, "");
           onChange({