+++ packages/spor-react/src/input/Select.tsx
@@ -85,13 +85,18 @@ export const Select = React.forwardRef<HTMLDivElement, SelectProps>(
           positioning={{ sameWidth: true, ...positioning }}
           variant={variant}
           css={styles.root}
-          position={"relative"}
+          position="relative"
         >
           <SelectTrigger data-attachable>
             <SelectValueText withPlaceholder={label ? true : false} />
           </SelectTrigger>
           {label && <SelectLabel css={styles.label}>{label}</SelectLabel>}
-          <SelectContent css={styles.selectContent}>{children}</SelectContent>
+          <SelectContent
+            css={styles.selectContent}
+            portalled={typeof document !== "undefined"}
+          >
+            {children}
+          </SelectContent>
         </ChakraSelect.Root>
       </Field>
     );
@@ -247,7 +252,7 @@ export const SelectValueText = React.forwardRef<
 
           collection: {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            stringifyItem: (arg0: any) => any;
+            stringifyItem: (argument0: any) => any;
           } /* Find a way to not use any */;
         }) => {
           const items = select.selectedItems;